schedule minor refactorings

+++ src/main/java/com/softserve/edu/manager/ScheduleRowsManager.java
@@ -28,7 +28,7 @@ public interface ScheduleRowsManager {
      *
      * @return Map<Long ,   String>.
      */
-    Map<Long, String> getWeekHead();
+    Map<Byte, String> getWeekHead();
 
     void setCalendar(Calendar calendar);
 }
